+++ src/mesido/esdl/esdl_heat_model.py
@@ -554,7 +554,7 @@ class AssetToHeatComponent(_AssetToComponentBase):
         # The asset attribute "dischargeEfficiency" represents the fraction of stored heat
         # that is lost per day. If this attribute is not provided in esdl, a default heat loss rate
         # of 1% per day is used. The value is converted to a per‑second loss factor.
-        heat_loss_efficiency = (
+        heat_loss_coefficient = (
             asset.attributes.get("dischargeEfficiency") / (24.0 * 3600.0)
             if asset.attributes.get("dischargeEfficiency")
             else 0.01 / (24.0 * 3600.0)
@@ -565,7 +565,7 @@ class AssetToHeatComponent(_AssetToComponentBase):
         modifiers = dict(
             height=r,
             radius=r,
-            heat_loss_efficiency=heat_loss_efficiency,
+            heat_loss_coefficient=heat_loss_coefficient,
             min_fraction_tank_volume=min_fraction_tank_volume,
             Stored_heat=dict(min=min_heat, max=max_heat),
             Heat_buffer=dict(min=-hfr_discharge_max, max=hfr_charge_max),