+++ tests/test_max_size_and_optional_assets.py
@@ -30,7 +30,6 @@ class TestMaxSizeAggregationCount(TestCase):
         assets should not be placed by the optimizer because of heat losses.
 
         Checks:
-        - Check the buffer heat loss calculation
         - Check that source 1 is utilized and also placed
         - Check that source 2 is utilized and placed
         - Check that the geothermal source is not placed
@@ -60,18 +59,6 @@ class TestMaxSizeAggregationCount(TestCase):
 
         results = solution.extract_results()
         parameters = solution.parameters(0)
-
-        # Test that if heat_loss_efficiency is incorporated into heat loss calculation of buffer
-        esdl_asset = solution.esdl_assets[solution.esdl_asset_name_to_id_map["HeatStorage_74c1"]]
-        np.testing.assert_allclose(
-            parameters["HeatStorage_74c1.heat_loss_efficiency"],
-            esdl_asset.attributes["dischargeEfficiency"] / (24.0 * 3600.0),
-        )
-        np.testing.assert_allclose(
-            results["HeatStorage_74c1.Heat_loss"],
-            results["HeatStorage_74c1.Stored_heat"]
-            * parameters["HeatStorage_74c1.heat_loss_efficiency"],
-        )
 
         # Producer 1 and geothermal source should not produce due to higher cost
         # Producer 2 should produce