+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Maximum profile constraint for PV asset is considered in PV sizing
 - Cater for input via esdl constraints to specify the upper limit for OPTIONAL assets in DTK
 - Initial implementation of adaptable pipe DN lower limit per pipe
-- heat_loss_efficiency of HeatStorage is read from dischargeEfficiency attribute in esdl
+- DischargeEfficiency is parsed from HeatStorage assets in ESDL to a heat loss coefficient
 
 
 ## Changed